// web/src/lib/types/SearchHit.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { NoteId } from "./NoteId";

/**
 * A search result hit
 */
export type SearchHit = { note_id: NoteId, title: string, snippet: string, rank: number, tags: Array<string>, path: string, };